Inlcude reinitialozation

+++ props_12.5keur.sh
@@ -78,9 +78,6 @@ if [ ! -f ${genos}/EURn${target}.bed ]
         $plink --bfile ${genos}/EUR --bmerge ${genos}/${target} --keep-allele-order --allow-no-sex --extract merged.totalsnps --make-bed --out ${genos}/EURn${target} --threads ${cpus} --memory $mem
 fi
 all=${genos}/EURn${target}
-#make train subset
-cat ${target}.train EUR.train > train.keep
-
 #if [ ! -f train.bed  ]; then
 #    $plink --bfile ${all}  --keep train.keep --keep-allele-order --allow-no-sex --make-bed --out train --memory $mem
 #fi
@@ -89,27 +86,41 @@ cat ${target}.train EUR.train > train.keep
 #if [ ! -f train.pca  ]; then
 #    python3 ${code}/skpca.py -b ${all} -t ${cpus} -m ${mem} -c 4
 #fi
-cat train.keep initial.keep | sort | uniq > train.txt
-$plink --bfile ${all} --keep train.txt --keep-allele-order --allow-no-sex --pca 4 --out train --threads ${cpus} --memory $mem
+#make train subset
+if [ ! -f train.txt ]
+    then
+        cat ${target}.train EUR.train > train.keep
+        cat train.keep initial.keep | sort | uniq > train.txt
+fi
+
+if [ ! -f train.pca  ]
+    then
+        $plink --bfile ${all} --keep train.txt --keep-allele-order --allow-no-sex --pca 4 --out train --threads ${cpus} --memory $mem
+fi
 
 step=$(( sample/10 ))
 
 # do Original
 prop=NONE
 const=NONE
-
+sequence=`seq 0 $step $sample`
 echo -e "\n\nStarting Original"
 if [ -f constant.tsv ]
     then
-        const1=(`tail -n 1 constant.tsv`)
-        const=${const1[0]}
+        pre=`cut -f1 constant.tsv`
+        sequ=`echo ${pre[@]} ${sequence[@]}| tr ' ' '\n'| sort| uniq -u`
+        # const1=(`tail -n 1 constant.tsv`)
+        # const=${const1[0]}
+    else
+        sequ=$sequence
 fi
-if [ -f proportions.tsv ]
-    then
-        prop1=(`tail -n 1 proportions.tsv`)
-        prop=${prop1[0]}
-fi
-for i in `seq 0 $step $sample`
+# if [ -f proportions.tsv ]
+#    then
+#        prop1=(`tail -n 1 proportions.tsv`)
+#        prop=${prop1[0]}
+# fi
+
+for i in sequ
 do
     eur=$(( sample - i ))
     echo -e "\n\nProcesing $eur european and $i $target"
@@ -124,15 +135,15 @@ do
             head -n $i ${target}.train >> constant_${i}.keep
     fi
     # Compute sumstats and clump for proportions
-    #${i}.eigenvec
+    # ${i}.eigenvec
     if [ ! -f props_${i}.clumped ]
         then
-            #$plink --bfile ${all} --keep ${i}.keep --keep-allele-order --allow-no-sex --pca 4 --out props_${i} --threads ${cpus} --memory $mem
-            #python3 ${code}/skpca.py -b ${all} -t ${cpus} -m ${mem} -c 4 --keep ${i}.keep -p 3
+            # $plink --bfile ${all} --keep ${i}.keep --keep-allele-order --allow-no-sex --pca 4 --out props_${i} --threads ${cpus} --memory $mem
+            # python3 ${code}/skpca.py -b ${all} -t ${cpus} -m ${mem} -c 4 --keep ${i}.keep -p 3
             $plink --bfile ${all} --keep ${i}.keep --keep-allele-order --allow-no-sex --linear hide-covar --pheno train.pheno --covar train.eigenvec --out props_${i} --threads ${cpus} --memory $mem
             $plink --bfile ${all} --keep ${i}.keep --keep-allele-order --allow-no-sex --clump props_${i}.assoc.linear --pheno train.pheno --out props_${i} --memory $mem
             # Do the constant estimations
-            #$plink --bfile ${all} --keep ${i}.keep --keep-allele-order --allow-no-sex --pca 4 --out constant_${i} --threads ${cpus} --memory $mem
+            # $plink --bfile ${all} --keep ${i}.keep --keep-allele-order --allow-no-sex --pca 4 --out constant_${i} --threads ${cpus} --memory $mem
             $plink --bfile ${all} --keep constant_${i}.keep --keep-allele-order --allow-no-sex --linear hide-covar --pheno train.pheno --covar train.eigenvec --out constant_${i} --threads ${cpus} --memory $mem
             $plink --bfile ${all} --keep constant_${i}.keep --keep-allele-order --allow-no-sex --clump constant_${i}.assoc.linear --pheno train.pheno --out constant_${i} --memory $mem
     fi
@@ -175,72 +186,76 @@ done
 # constant initial source add mixing
 if [ -f init12k.tsv ]
     then
-        lines=`wc -l init12k.tsv`
-        echo -e "\n\ninit12k has been done, wc is $lines"
+        pre=`cut -f1 init12k.tsv`
+        sequ=`echo ${pre[@]} ${sequence[@]}| tr ' ' '\n'| sort| uniq -u`
     else
-        echo -e "\n\nStarting constant initial source add mixing"
-        for i in `seq 0 $step $sample`
-        do
-            eur=$(( sample - i ))
-            echo -e "\n\nProcesing $eur european and $i $target with start of $init"
-            cat initial.keep > init_${i}.keep
-            if [[ ! $i = 0 ]]; then head -n $i ${target}.train >> init_${i}.keep; fi
-            if [[ ! $eur = 0  ]]; then head -n $eur EUR.train >> init_${i}.keep; fi
-            # compute pca for this subset
-            #$plink --bfile ${all} --keep init_${i}.keep --keep-allele-order --allow-no-sex --pca 4 --out init_${i} --threads ${cpus} --memory $mem
-            # Perform associations and clumping
-            $plink --bfile ${all} --keep init_${i}.keep --keep-allele-order --allow-no-sex --linear hide-covar --pheno train.pheno --covar train.eigenvec --out init_${i} --threads ${cpus} --memory $mem
-            $plink --bfile ${all} --keep init_${i}.keep --keep-allele-order --allow-no-sex --clump init_${i}.assoc.linear --pheno train.pheno --out init_${i} --memory $mem
-            grep -w "$(awk -F' ' '{if (NR!=1) { print $3 }}' init_${i}.clumped)" init_${i}.assoc.linear > init_${i}.myscore
-            # Score original
-            $plink --bfile ${target}_test --score init_${i}.myscore 2 4 7 sum center --pheno train.pheno --keep-allele-order --allow-no-sex --out ${target}_init_${i} --threads ${cpus} --memory $mem
-            outp ${target}_init_${i}.profile ${target} init12k.tsv
-            $plink --bfile EUR_test --score init_${i}.myscore 2 4 7 sum center --pheno train.pheno --keep-allele-order --allow-no-sex --out EUR_init_${i} --threads ${cpus} --memory $mem
-            outp EUR_init_${i}.profile EUR init12k.tsv
-            $plink --bfile ${genos}/AFR --score init_${i}.myscore 2 4 7 sum center --pheno AFR.pheno --keep-allele-order --allow-no-sex --out AFR_init_${i} --threads ${cpus} --memory $mem
-            outp AFR_init_${i}.profile AFR init12k.tsv
-
-            # python3 ${code}/simple_score.py -b ${target}_test -c ${i}.clumped -s ${i}.assoc.linear -t ${cpus} -p train.pheno -l ${target} -m $membytes -P init12k
-            # python3 ${code}/simple_score.py -b EUR_test -c ${i}.clumped -s ${i}.assoc.linear -t ${cpus} -p train.pheno -l EUR -m $membytes -P init12k
-            # python3 ${code}/simple_score.py -b ${genos}/AFR -c ${i}.clumped -s ${i}.assoc.linear -t ${cpus} -p AFR.pheno -l AFR -m $membytes -P init12k
-        done
+        sequ=${sequence}
 fi
+echo -e "\n\nStarting constant initial source add mixing"
+for i in sequ
+do
+    eur=$(( sample - i ))
+    echo -e "\n\nProcesing $eur european and $i $target with start of $init"
+    cat initial.keep > init_${i}.keep
+    if [[ ! $i = 0 ]]; then head -n $i ${target}.train >> init_${i}.keep; fi
+    if [[ ! $eur = 0  ]]; then head -n $eur EUR.train >> init_${i}.keep; fi
+    # compute pca for this subset
+    #$plink --bfile ${all} --keep init_${i}.keep --keep-allele-order --allow-no-sex --pca 4 --out init_${i} --threads ${cpus} --memory $mem
+    # Perform associations and clumping
+    $plink --bfile ${all} --keep init_${i}.keep --keep-allele-order --allow-no-sex --linear hide-covar --pheno train.pheno --covar train.eigenvec --out init_${i} --threads ${cpus} --memory $mem
+    $plink --bfile ${all} --keep init_${i}.keep --keep-allele-order --allow-no-sex --clump init_${i}.assoc.linear --pheno train.pheno --out init_${i} --memory $mem
+    grep -w "$(awk -F' ' '{if (NR!=1) { print $3 }}' init_${i}.clumped)" init_${i}.assoc.linear > init_${i}.myscore
+    # Score original
+    $plink --bfile ${target}_test --score init_${i}.myscore 2 4 7 sum center --pheno train.pheno --keep-allele-order --allow-no-sex --out ${target}_init_${i} --threads ${cpus} --memory $mem
+    outp ${target}_init_${i}.profile ${target} init12k.tsv
+    $plink --bfile EUR_test --score init_${i}.myscore 2 4 7 sum center --pheno train.pheno --keep-allele-order --allow-no-sex --out EUR_init_${i} --threads ${cpus} --memory $mem
+    outp EUR_init_${i}.profile EUR init12k.tsv
+    $plink --bfile ${genos}/AFR --score init_${i}.myscore 2 4 7 sum center --pheno AFR.pheno --keep-allele-order --allow-no-sex --out AFR_init_${i} --threads ${cpus} --memory $mem
+    outp AFR_init_${i}.profile AFR init12k.tsv
+
+    # python3 ${code}/simple_score.py -b ${target}_test -c ${i}.clumped -s ${i}.assoc.linear -t ${cpus} -p train.pheno -l ${target} -m $membytes -P init12k
+    # python3 ${code}/simple_score.py -b EUR_test -c ${i}.clumped -s ${i}.assoc.linear -t ${cpus} -p train.pheno -l EUR -m $membytes -P init12k
+    # python3 ${code}/simple_score.py -b ${genos}/AFR -c ${i}.clumped -s ${i}.assoc.linear -t ${cpus} -p AFR.pheno -l AFR -m $membytes -P init12k
+done
+
 # do the cost derived
+sequence=`seq 0 10`
 if [ -f cost.tsv ]
     then
-        lines2=`wc -l cost.tsv`
-        echo -e "\n\nCost has been done, wc is $lines2"
+        pre=`cut -f1 cost.tsv`
+        sequ=`echo ${pre[@]} ${sequence[@]}| tr ' ' '\n'| sort| uniq -u`
     else
-        echo -e "\n\nStarting cost"
-        for j in `seq 0 10`
-        do
-            eu=`bc <<< "scale = 1; $j/10"`
-            ad=`bc <<< "scale = 1; 1 - ($j/10)"`
-            n=`bc <<< "scale = 1; $sample / (($ad * 2) + $eu)"`
-            n=`bc <<< "$n/1"`
-            eu=`bc <<< "($n * $eu)/1"`
-            ad=`bc <<< "($n * $ad)/1"`
-            echo -e "\n\nProcesing $eu european and $ad admixed"
-            if [[ ! $ad = 0  ]]; then
-                sort -R ${target}.train| head -n $ad > frac_${j}.keep
-            fi
-            if [[ ! $eu = 0  ]]; then
-                sort -R EUR.train| head -n $eu >> frac_${j}.keep
-            fi
-                # Perform associations and clumping
-            #$plink --bfile ${all} --keep frac_${j}.keep --keep-allele-order --allow-no-sex --pca 4 --out frac_${j} --threads ${cpus} --memory $mem
-            $plink --bfile ${all} --keep frac_${j}.keep --keep-allele-order --allow-no-sex --linear hide-covar --pheno train.pheno --covar train.eigenvec --out cost_${j} --threads ${cpus} --memory $mem
-            $plink --bfile ${all} --keep frac_${j}.keep --keep-allele-order --allow-no-sex --clump cost_${j}.assoc.linear --pheno train.pheno --out cost_${j} --memory $mem
-            # Score cost
-            grep -w "$(awk -F' ' '{if (NR!=1) { print $3 }}' cost_${j}.clumped)" cost_${j}.assoc.linear > cost_${j}.myscore
-            $plink --bfile ${target}_test --score cost_${j}.myscore 2 4 7 sum center --pheno train.pheno --keep-allele-order --allow-no-sex --out ${target}_${j}_cost --threads ${cpus} --memory $mem
-            outp ${target}_${j}_cost.profile ${target} cost.tsv
-            $plink --bfile EUR_test --score cost_${j}.myscore 2 4 7 sum center --pheno train.pheno --keep-allele-order --allow-no-sex --out EUR_${j}_cost --threads ${cpus} --memory $mem
-            outp EUR_${j}_cost.profile EUR cost.tsv
-            $plink --bfile ${genos}/AFR --score cost_${j}.myscore 2 4 7 sum center --pheno AFR.pheno --keep-allele-order --allow-no-sex --out AFR_${j}_cost --threads ${cpus} --memory $mem
-            outp AFR_${j}_cost.profile AFR cost.tsv
-            # python3 ${code}/simple_score.py -b ${target}_test -c cost_${j}.clumped -s cost_${j}.assoc.linear -t ${cpus} -p train.pheno -l ${target} -m $membytes -P cost
-            # python3 ${code}/simple_score.py -b EUR_test -c cost_${j}.clumped -s cost_${j}.assoc.linear -t ${cpus} -p train.pheno -l EUR -m $membytes -P cost
-            # python3 ${code}/simple_score.py -b ${genos}/AFR -c cost_${j}.clumped -s cost_${j}.assoc.linear -t ${cpus} -p AFR.pheno -l AFR -m $membytes -P cost
-        done
+        sequ=${sequence}
 fi
+echo -e "\n\nStarting cost"
+for j in sequ
+do
+    eu=`bc <<< "scale = 1; $j/10"`
+    ad=`bc <<< "scale = 1; 1 - ($j/10)"`
+    n=`bc <<< "scale = 1; $sample / (($ad * 2) + $eu)"`
+    n=`bc <<< "$n/1"`
+    eu=`bc <<< "($n * $eu)/1"`
+    ad=`bc <<< "($n * $ad)/1"`
+    echo -e "\n\nProcesing $eu european and $ad admixed"
+    if [[ ! $ad = 0  ]]; then
+        sort -R ${target}.train| head -n $ad > frac_${j}.keep
+    fi
+    if [[ ! $eu = 0  ]]; then
+        sort -R EUR.train| head -n $eu >> frac_${j}.keep
+    fi
+        # Perform associations and clumping
+    # $plink --bfile ${all} --keep frac_${j}.keep --keep-allele-order --allow-no-sex --pca 4 --out frac_${j} --threads ${cpus} --memory $mem
+    $plink --bfile ${all} --keep frac_${j}.keep --keep-allele-order --allow-no-sex --linear hide-covar --pheno train.pheno --covar train.eigenvec --out cost_${j} --threads ${cpus} --memory $mem
+    $plink --bfile ${all} --keep frac_${j}.keep --keep-allele-order --allow-no-sex --clump cost_${j}.assoc.linear --pheno train.pheno --out cost_${j} --memory $mem
+    # Score cost
+    grep -w "$(awk -F' ' '{if (NR!=1) { print $3 }}' cost_${j}.clumped)" cost_${j}.assoc.linear > cost_${j}.myscore
+    $plink --bfile ${target}_test --score cost_${j}.myscore 2 4 7 sum center --pheno train.pheno --keep-allele-order --allow-no-sex --out ${target}_${j}_cost --threads ${cpus} --memory $mem
+    outp ${target}_${j}_cost.profile ${target} cost.tsv
+    $plink --bfile EUR_test --score cost_${j}.myscore 2 4 7 sum center --pheno train.pheno --keep-allele-order --allow-no-sex --out EUR_${j}_cost --threads ${cpus} --memory $mem
+    outp EUR_${j}_cost.profile EUR cost.tsv
+    $plink --bfile ${genos}/AFR --score cost_${j}.myscore 2 4 7 sum center --pheno AFR.pheno --keep-allele-order --allow-no-sex --out AFR_${j}_cost --threads ${cpus} --memory $mem
+    outp AFR_${j}_cost.profile AFR cost.tsv
+    # python3 ${code}/simple_score.py -b ${target}_test -c cost_${j}.clumped -s cost_${j}.assoc.linear -t ${cpus} -p train.pheno -l ${target} -m $membytes -P cost
+    # python3 ${code}/simple_score.py -b EUR_test -c cost_${j}.clumped -s cost_${j}.assoc.linear -t ${cpus} -p train.pheno -l EUR -m $membytes -P cost
+    # python3 ${code}/simple_score.py -b ${genos}/AFR -c cost_${j}.clumped -s cost_${j}.assoc.linear -t ${cpus} -p AFR.pheno -l AFR -m $membytes -P cost
+done
